% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-abundanceperquad.R
\name{abundance}
\alias{abundance}
\title{Internal.}
\usage{
abundance(censdata, type = "abund", alivecode = c("A"), mindbh = NULL,
  dbhunit = "mm", split1 = NULL, split2 = NULL)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundanceperquad}},
  \code{\link{basum}}, \code{\link{ba}},
  \code{\link{fill.dimension}},
  \code{\link{gxgy.to.index}},
  \code{\link{rowcol.to.index}}
}
\keyword{internal}
