% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-abundanceperquad.R
\name{abundanceperquad}
\alias{abundanceperquad}
\title{Abundance, basal area, or agb of every species by quadrat.}
\usage{
abundanceperquad(censdata, mindbh = 10, plotdim = c(1000, 500),
  gridsize = 100, type = "abund", dbhunit = "mm")
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{gridsize}{Side of the square quadrate.}

\item{type}{Either 'abund' (default) for abundance, 'ba' for basal area, or
'agb'.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}
}
\description{
Finds abundance, basal area, or agb of every species per square quadrat of
any size; plotdim is the x dimension then y dimension of the plot and must be
set correctly; gridsize is the quadrat dimension. The plot is divided into a
checkerboard of non-overlapping, space-filling squares.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundance}},
  \code{\link{basum}}, \code{\link{ba}},
  \code{\link{fill.dimension}},
  \code{\link{gxgy.to.index}},
  \code{\link{rowcol.to.index}}
}
\keyword{internal}
