% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df.krig_lst}
\alias{as_df.krig_lst}
\title{Dataframe the output of \code{krig_lst()}.}
\usage{
\method{as_df}{krig_lst}(krig_lst, name = "var", item = "df")
}
\arguments{
\item{krig_lst}{The output of \code{\link[=krig_lst]{krig_lst()}}.}

\item{name}{Name for the column to hold soil variable-names.}

\item{item}{Character string; either "df" or "df.poly".}
}
\value{
A dataframe.
}
\description{
Dataframe the output of \code{krig_lst()}.
}
\examples{
vars <- c("c", "p")
krig_lst <- krig_lst(vars, soil_fake, quiet = TRUE)
as_df(krig_lst)
}
