% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df}
\alias{as_df}
\title{Restructure data as a dataframe.}
\usage{
as_df(.x, ...)
}
\arguments{
\item{.x}{An fgeo object of supported class.}

\item{...}{Other arguments passed to methods.}
}
\value{
A dataframe.
}
\description{
The goal of this (generic) function (with methods for multiple fgeo classes)
is to produce a dataframe that helps you to work fluently with other general
purpose tools such as \strong{dplyr} and \strong{ggplot2}.
}
\examples{
# Method for class "krig_lst"
vars <- c("c", "p")
krig_lst <- krig_lst(soil_fake, vars, quiet = TRUE)
as_df(krig_lst)
}
