% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df_tt}
\alias{as_df_tt}
\alias{as_df.tt_one}
\alias{as_df.tt_lst}
\title{Dataframe objects of class tt_*.}
\usage{
\method{as_df}{tt_one}(.x, ...)

\method{as_df}{tt_lst}(.x, ...)
}
\arguments{
\item{.x}{An object of class tt_*.}

\item{...}{Other arguments passed to \code{\link[=as_df]{as_df()}}.}
}
\value{
A dataframe.
}
\description{
Dataframe objects of class tt_*.
}
\examples{
cns <- fgeo.habitat::luquillo_top3_sp
spp <- unique(cns$sp)
hab <- luquillo_habitat

tt_lst <- tt_test_lst(spp, cns, hab)
tt_df <- as_df(tt_lst)
head(tt_df)

tail(tt_df)
}
