% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df_ttl}
\alias{as_df_ttl}
\title{Dataframe objects of class tt_*.}
\arguments{
\item{.x}{An object of class tt_*.}

\item{...}{Other arguments passed to \code{\link[=as_df]{as_df()}}.}
}
\value{
A dataframe.
}
\description{
Dataframe objects of class tt_*.
}
\examples{
# Class tt_lst
cns <- fgeo.habitat::luquillo_top3_sp
spp <- unique(cns$sp)
hab <- luquillo_habitat

tt_lst <- tt_test_lst(spp, cns, hab)
tt_df <- as_df(tt_lst)
head(tt_df)

tail(tt_df)

# Class tt

pdim <- c(320, 500)
gsz <- 20
abnd <- abund_index(cns, pdim, gsz)
spp1 <- spp[[1]]

tt <- tt_test(
  species = spp1,
  hab.index20 = hab,
  allabund20 = abnd,
  plotdim = pdim,
  gridsize = gsz
)

tt_df <- as_df(tt)
head(tt_df)
tail(tt_df)
}
