% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_krig.R
\name{as_krig}
\alias{as_krig}
\alias{as_krig_lst}
\alias{as_krig_lst.krig_lst}
\alias{as_krig_lst.list}
\alias{as_krig_lst.default}
\title{Coerce to class krig_lst}
\usage{
as_krig(.x, ...)

as_krig_lst(.x, ...)

\method{as_krig_lst}{krig_lst}(.x, ...)

\method{as_krig_lst}{list}(.x, ...)

\method{as_krig_lst}{default}(.x, ...)
}
\arguments{
\item{.x}{Output formatted as that of \code{krig_lst()}.}

\item{...}{Other arguments passed to methods.}
}
\value{
An object of class krig.
}
\description{
Usually you would use this function to coerce the output of \code{GetKrigedSoil()}
and because you want to use the methods available for objects of class krig
or krig_lst.
}
\examples{
out <- GetKrigedSoil(soil_fake, "c")
summary(out)
summary(as_krig(out))

out_lst <- lapply(c("c", "p"), function(var) GetKrigedSoil(soil_fake, var))
head(to_df(as_krig_lst(out_lst)))
}
\seealso{
\code{\link[=summary.krig]{summary.krig()}}, \code{\link[=to_df.krig_lst]{to_df.krig_lst()}}.
}
