% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tt.R
\name{as_tt}
\alias{as_tt}
\alias{as_tt_lst}
\title{Coerce to class tt and tt_list.}
\usage{
as_tt(.x, ...)

as_tt_lst(.x, ...)
}
\arguments{
\item{.x}{Output structured as that of \code{\link[=tt_test]{tt_test()}} or \code{\link[=tt_test_lst]{tt_test_lst()}}.}

\item{...}{Other arguments passed to methods.}
}
\value{
An object of class tt or tt_lst.
}
\description{
Coerce to class tt and tt_list.
}
\examples{
# For easier data wranging
library(dplyr)

habitat <- luquillo_habitat
census <- luquillo_top3_sp

# Pick alive trees, of 10 mm or more
pick <- filter(census, status == "A", dbh >= 10)
# Pick sufficiently abundant trees
pick <- add_count(pick, sp)
pick <- filter(pick, n > 50)

species <- unique(pick$sp)
plotdim <- c(320, 500)
gridsize <- 20
abundance <- abund_index(pick, plotdim, gridsize)

tt_mat <- torusonesp.all(species[[1]],
  hab.index20 = habitat,
  allabund20 = abundance,
  plotdim = plotdim,
  gridsize = gridsize
)

head(as_df(as_tt(tt_mat)))

# Iterate over multiple species
tt_lst <- lapply(species, tt_test, habitat, abundance, plotdim, gridsize)
head(as_df(as_tt_lst(tt_lst)))
}
\seealso{
\code{\link[=as_df]{as_df()}}.
}
