% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-abundanceperquad.R
\name{basum}
\alias{basum}
\title{Internal.}
\usage{
basum(dbh, mindbh = 10, dbhunit = "mm")
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundanceperquad}},
  \code{\link{abundance}}, \code{\link{ba}},
  \code{\link{fill.dimension}},
  \code{\link{gxgy.to.index}},
  \code{\link{rowcol.to.index}}
}
\keyword{internal}
