% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.habitat-package.R
\docType{package}
\name{fgeo.habitat-package}
\alias{fgeo.habitat}
\alias{fgeo.habitat-package}
\title{fgeo.habitat: Analize Soils and Tree-Habitat Data}
\description{
Analize soils and tree-habitat data.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Authors:
\itemize{
  \item Graham Zemunik \email{grah.zem@@gmail.com}
}

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
