% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-abundanceperquad.R
\name{fill.dimension}
\alias{fill.dimension}
\title{Internal.}
\usage{
fill.dimension(dataarray, class1, class2, fill = 0)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundanceperquad}},
  \code{\link{abundance}}, \code{\link{basum}},
  \code{\link{ba}}, \code{\link{gxgy.to.index}},
  \code{\link{rowcol.to.index}}
}
\keyword{internal}
