% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKrigedSoil.R
\name{GetAutomatedKrigeParams}
\alias{GetAutomatedKrigeParams}
\title{Find the "best" variogram parameters for a given geodata object.}
\usage{
GetAutomatedKrigeParams(geod, trend = "cte", breaks = ExpList(2, 320, 30))
}
\arguments{
\item{geod, trend, breaks}{Passed to \code{\link[geoR:variog]{geoR::variog()}}.}
}
\value{
A list of the best fitted variogram parameters:
\itemize{
\item nugget: FIXME.
\item sill: FIXME.
\item range: FIXME.
\item kappa: FIXME.
\item model: FIXME.
\item minVM: The minimum fit error.
\item vg: The variogram.
}
}
\description{
Uses the \code{\link[geoR:variofit]{geoR::variofit()}} with a range of variogram models.
}
\examples{
\dontrun{
# See examples in geoR::variog().
}
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
