% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKrigedSoil.R
\name{GetKrigedSoil}
\alias{GetKrigedSoil}
\title{Krige soil data following the methodology of the John et al. (2007).}
\usage{
GetKrigedSoil(df.soil, var = "P", gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = ExpList(2, 320, 30),
  useKsLine = TRUE)
}
\arguments{
\item{df.soil}{The data frame with the points, coords specified in the
columns `gx`, `gy`.}

\item{var}{The variable/column in `df.soil` to krige.}

\item{gridSize}{Points are kriged to the centre points of a grid of this
size.}

\item{krigeParams}{If you want to pass specified kriging parameters; see
[GetAutomatedKrigeParams()] for each parameter.}

\item{xSize, ySize}{X and Y size/length of the plot.}

\item{breaks}{Breaks/intervals used to calculate the semivariogram, which
only happens if `krigeParams = NULL` (default).}

\item{useKsLine}{See above.}
}
\value{
A list with the following items:
  * `df`: Data frame of kriged values (column z) at each grid point (x, y).
  * `df.poly`: Data frame of the polynomial surface fitted to the raw data.
  * `lambda`: The "lambda" value used in the Box-Cox transform of the raw
    data.
  * `vg`: The variogram parameters used for the kriging.
  * `vm`: Minimum loss value returned from [geoR::variofit()].
}
\description{
If called without inputs, `GetKrigedSoil()` calculates a "best" semivariogram
to use and krige the soil data.
}
\details{
By default the best kriging parameters (found via variogram functions in the
__geoR__ package) are used, but specified parameters can be used via the
`krigeParams` argument geoR has two main kriging functions: [geoR::ksline()]
and [geoR::krige.conv()]. The argument `useKsLine` specifies whether to use
the [geoR::ksline()] function or not.
}
\examples{
\dontrun{
library(ggplot2)

# Use randomized data set for examples (invalid for reaserch).
df <- krig::soil_random
str(df)

# Krige with automated parameters
kriged <- GetKrigedSoil( df, var="M3Al" )
str(kriged)

# Krige with specified parameters - these example params are rather arbitrary
# but are loosely based on what was chosen from the automated kriging
params <- list( model="circular", range=100, nugget=1000, sill=46000, kappa=0.5 )
kriged.2 <- GetKrigedSoil( df, var="M3Al", krigeParams=params )

# Have a look at the differences
g <- ggplot( kriged$df, aes( x=x, y=y, fill=z ) )
g <- g + geom_tile() + coord_equal()
g2 <- ggplot( kriged.2$df, aes( x=x, y=y, fill=z ) )
g2 <- g2 + geom_tile() + coord_equal()
g
g2
}
}
\seealso{
[geoR::variofit()], [geoR::variog()], [geoR::as.geodata()],
  [geoR::ksline()], [geoR::krige.conv()], [geoR::krige.control()].
}
