% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig-package.R
\docType{package}
\name{krig-package}
\alias{krig}
\alias{krig-package}
\title{krig: Krig Soil Data}
\description{
Krig soil data as per John et al. (2007).
}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Authors:
\itemize{
  \item Graham Zemunik \email{grah.zem@@gmail.com}
}

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
