% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig.R
\name{krig_auto_params}
\alias{krig_auto_params}
\title{Find "best" variogram parameters.}
\usage{
krig_auto_params(geodata, trend = "cte", breaks = krig_breaks(2, 320, 30))
}
\value{
A list of the best fitted variogram parameters. The following
descriptioin is adapted from \code{\link[geoR:variog]{geoR::variog()}} -- which you should see
for more details:
\itemize{
\item nugget: value of the nugget parameter. An estimated value if
\code{fix.nugget = FALSE} or a fixed value if \code{fix.nugget = TRUE}.
\item sill and range: First and second elements of cov.pars` -- a two elements
vector with estimated values of the covariance parameters sigma^2 and phi,
respectively.
\item kappa: Fixed value of the smoothness parameter.
\item model: Name of the correlation function (see\code{cov.model}).
\item minVM: The minimum fit error.
\item vg: The variogram.
}
}
\description{
Find the "best" variogram parameters for a given geodata object.
}
\details{
The default breaks argument is set to have more points where the curve rises
the most and exponentially and fewer at large distances. This means that the
curve fitting is not overly biased by points beyond the effective maximum
range.

Several different models are tested; the one with the lowest least
squares error is chosen.
}
\seealso{
\link[geoR:variog]{geoR::variog}.
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
