% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_versions_of_geoR.R
\name{krig_ksline}
\alias{krig_ksline}
\title{Version of geoR::ksline that avoids a warning.}
\usage{
krig_ksline(geodata, coords = geodata$coords, data = geodata$data,
  locations, borders = NULL, cov.model = "matern",
  cov.pars = stop("covariance parameters (sigmasq and phi) needed"),
  kappa = 0.5, nugget = 0, micro.scale = 0, lambda = 1, m0 = "ok",
  nwin = "full", n.samples.backtransform = 500, trend = 1, d = 2,
  ktedata = NULL, ktelocations = NULL, aniso.pars = NULL,
  signal = FALSE, dist.epsilon = 1e-10, messages)
}
\description{
See https://github.com/forestgeo/krig/issues/9.
}
\seealso{
\code{\link[geoR:ksline]{geoR::ksline()}}
}
