% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luquillo}
\alias{luquillo}
\alias{luquillo_habitat}
\alias{luquillo_elevation}
\alias{luquillo_tree6_random}
\alias{luquillo_stem6_random}
\title{Datasets from \strong{fgeo.data}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 3 columns.}
\usage{
luquillo_habitat

luquillo_elevation

luquillo_tree6_random

luquillo_stem6_random
}
\description{
Datasets from \strong{fgeo.data}.
}
\examples{
str(luquillo_habitat, give.attr = FALSE)
str(luquillo_elevation, give.attr = FALSE)
str(luquillo_tree6_random, give.attr = FALSE)
str(luquillo_stem6_random, give.attr = FALSE)
}
\seealso{
\code{fgeo.data::luquillo_habitat}, \code{fgeo.data::luquillo_elevation},
\code{fgeo.data::luquillo_tree6_random}, \code{fgeo.data::luquillo_stem6_random}.
}
\keyword{datasets}
