% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_enframe.R
\name{mat_enframe}
\alias{mat_enframe}
\alias{mat_enframe_ls}
\title{Convert matrices to data frames.}
\usage{
mat_enframe(...)

mat_enframe_ls(mat, rownm = "rownames", colnm = "colnames",
  value = "value")
}
\arguments{
\item{...}{Arguments passed to mat_enframe_ls.}

\item{mat}{A matrix.}

\item{rownm, colnm, value}{Names of the columns that store the row names, the
column names and the values.}
}
\value{
\itemize{
\item \code{\link[=mat_enframe]{mat_enframe()}}: A dataframe.
\item \code{\link[=mat_enframe_ls]{mat_enframe_ls()}}: A list of dataframes.
}
}
\description{
These functions convert a matrix to three-column data frames, where columns
store row names, column names, and values of the original matrix.
\code{\link[=mat_enframe]{mat_enframe()}} outputs a long-format dataframe and \code{\link[=mat_enframe_ls]{mat_enframe_ls()}}
returns a list of dataframes where each element of the list maps to a column
of the original matrix.
}
\details{
These functions are a matrix implementation of the \code{enframe()} function of
the \strong{tibble} package.
}
\examples{
 mat <- matrix(1:6, 2, dimnames = list(LETTERS[1:2], letters[1:3]))
 mat
 
 mat_enframe_ls(mat)
 
 mat_enframe(mat)
 
 mat_enframe(mat, "metric", "sp")
 
 mat <- matrix(1:6, 2)
 mat
 
 mat_enframe(mat)
}
\seealso{
\code{enframe()} (\strong{tibble} package).
}
