% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.habitat-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{guess_plotdim}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fgeo.base}{\code{\link[fgeo.base]{guess_plotdim}}}
}}

