% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-abundanceperquad.R
\name{rowcol.to.index}
\alias{rowcol.to.index}
\title{Internal.}
\usage{
rowcol.to.index(rowno, colno, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundanceperquad}},
  \code{\link{abundance}}, \code{\link{basum}},
  \code{\link{ba}}, \code{\link{fill.dimension}},
  \code{\link{gxgy.to.index}}
}
\keyword{internal}
