% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_interpret.R
\name{summary.tt_lst}
\alias{summary.tt_lst}
\title{Interpret the output of \code{tt_test()}.}
\usage{
\method{summary}{tt_lst}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link[=tt_test]{tt_test()}}, i.e. an S3 object of class tt_lst.}

\item{...}{Other arguments passed to methods.}
}
\value{
A dataframe.
}
\description{
Interpret the output of \code{tt_test()}.
}
\author{
Daniel Zuleta.
}
