% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.krig_lst}
\alias{to_df.krig_lst}
\title{Dataframe the output of \code{krig_lst()}.}
\usage{
\method{to_df}{krig_lst}(.x, name = "var", item = "df", ...)
}
\arguments{
\item{.x}{The output of \code{\link[=krig_lst]{krig_lst()}}.}

\item{name}{Name for the column to hold soil variable-names.}

\item{item}{Character string; either "df" or "df.poly".}

\item{...}{Arguments passed on to \code{to_df}
\describe{
  \item{.x}{An fgeo object of supported class.}
}}
}
\value{
A dataframe.
}
\description{
Dataframe the output of \code{krig_lst()}.
}
\examples{
vars <- c("c", "p")
krig_lst <- krig_lst(soil_fake, vars, quiet = TRUE)
head(to_df(krig_lst))
}
