% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.tt}
\alias{to_df.tt}
\alias{to_df.tt_lst}
\title{Dataframe objects of class tt_*.}
\usage{
\method{to_df}{tt}(.x, ...)

\method{to_df}{tt_lst}(.x, ...)
}
\arguments{
\item{.x}{An object of class tt_*.}

\item{...}{Other arguments passed to \code{\link[=to_df]{to_df()}}.}
}
\value{
A dataframe.
}
\description{
Dataframe objects of class tt_*.
}
\examples{
# Class tt_lst
cns <- fgeo.habitat::luquillo_top3_sp
spp <- unique(cns$sp)
hab <- luquillo_habitat

tt_lst <- tt_test_lst(cns, spp, hab)
tt_df <- to_df(tt_lst)
head(tt_df)

tail(tt_df)

# Class tt

pdim <- c(320, 500)
gsz <- 20
abnd <- abund_index(cns, pdim, gsz)
spp1 <- spp[[1]]

tt <- tt_test(spp1, hab, abnd, pdim, gsz)
tt_df <- to_df(tt)
head(tt_df)

tail(tt_df)
}
