% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.tt_lst}
\alias{to_df.tt_lst}
\title{Dataframe the output of tt_test().}
\usage{
\method{to_df}{tt_lst}(.x, ...)
}
\arguments{
\item{.x}{An object of class tt_lst.}

\item{...}{Other arguments passed to \code{\link[=to_df]{to_df()}}.}
}
\value{
A dataframe.
}
\description{
Dataframe the output of tt_test().
}
\examples{
cns <- fgeo.habitat::luquillo_top3_sp
spp <- unique(cns$sp)
hab <- luquillo_habitat

tt_lst <- tt_test(cns, spp, hab)
tt_df <- to_df(tt_lst)
head(tt_df)

tail(tt_df)
}
