% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torusonesp.all.R
\name{torusonesp.all}
\alias{torusonesp.all}
\title{Torus Translation Test to determine habitat associations of tree species.}
\usage{
torusonesp.all(species, hab.index20, allabund20, plotdim, gridsize)
}
\arguments{
\item{species}{Character sting giving the name of one species.}

\item{hab.index20}{Object giving the habitat designation for each 20x20
quadrat.}

\item{allabund20}{Abundance per quadrat.}

\item{plotdim}{Plot dimensions.}

\item{gridsize}{Grid size. Make sure the gridsize entered here matches the
gridsize on which the habitats are defined and the abundances were
calculated}
}
\description{
Torus Translation Test (TT test) to determine habitat associations of tree
species. Based on code written by Kyle Harms.
}
\details{
You should only try to determine the habitat association for sufficiently
abundant species - in a 50-ha plot, a minimum abundance of 50 trees/species
has been used.
}
\author{
Sabrina Russo, Daniel Zuletta, Matteo Detto.
}
