% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_test.R
\name{tt_df}
\alias{tt_df}
\title{Gather the output of \code{tt_test_one()}.}
\usage{
tt_df(ttt)
}
\arguments{
\item{ttt}{Output of \code{tt_test_one()}; Either a single matrix or a list of
matrices.}
}
\value{
A dataframe.
}
\description{
Dataframe the output of \code{tt_test_one()}. For examples see \code{\link[=tt_test]{tt_test()}}.
}
