% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_test.R
\name{tt_test_lst}
\alias{tt_test_lst}
\title{#' @rdname tt_test}
\usage{
tt_test_lst(sp, census, habitat, plotdim = extract_plotdim(habitat),
  gridsize = extract_gridsize(habitat))
}
\description{
#' @rdname tt_test
}
