% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_test.R
\name{tt_test_ply}
\alias{tt_test_ply}
\title{Returns a list of results for each species.}
\usage{
tt_test_ply(sp, census, habitat, plotdim, gridsize)
}
\description{
Returns a list of results for each species.
}
\keyword{internal}
