% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKrigedSoil.R
\name{ExpList}
\alias{ExpList}
\title{List of n values which exponentially increases from first to last.}
\usage{
ExpList(first, last, n)
}
\arguments{
\item{first}{Xxx.}

\item{last}{Xxx.}

\item{n}{Xxx.}
}
\description{
List of n values which exponentially increases from first to last.
}
\examples{
ExpList(2, 320, 30)
}
