% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.habitat-package.R
\docType{package}
\name{fgeo.habitat-package}
\alias{fgeo.habitat}
\alias{fgeo.habitat-package}
\title{fgeo.habitat: Analize Soils and Tree-Habitat Data}
\description{
Analize soils and tree-habitat data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.habitat}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.habitat/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
