% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKrigedSoil.R
\name{GetKrigedSoil}
\alias{GetKrigedSoil}
\title{Krige soil data following the methodology of the John et al. (2007).}
\usage{
GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = ExpList(2, 320, 30),
  useKsLine = TRUE)
}
\arguments{
\item{df.soil}{The data frame with the points, coords specified in the
columns \code{gx}, \code{gy}.}

\item{var}{A text string giving the variable/column in \code{df.soil} to krige.}

\item{gridSize}{Points are kriged to the center points of a grid of this
size.}

\item{krigeParams}{If you want to pass specified kriging parameters; see
\code{\link[=GetAutomatedKrigeParams]{GetAutomatedKrigeParams()}} for each parameter.}

\item{xSize, ySize}{X and Y size/length of the plot.}

\item{breaks}{Breaks/intervals used to calculate the semivariogram, which
only happens if \code{krigeParams = NULL} (default).}

\item{useKsLine}{See above.}
}
\value{
A list with the following items:
\itemize{
\item \code{df}: Data frame of kriged values (column z) at each grid point (x, y).
\item \code{df.poly}: Data frame of the polynomial surface fitted to the raw data.
\item \code{lambda}: The "lambda" value used in the Box-Cox transform of the raw
data.
\item \code{vg}: A list giving the variogram parameters used for the kriging.
\item \code{vm}: Minimum loss value returned from \code{\link[geoR:variofit]{geoR::variofit()}}.
}
}
\description{
If called without inputs, \code{GetKrigedSoil()} calculates a "best" semivariogram
to use and krige the soil data. By default the best kriging parameters (found
via variogram functions in the \strong{geoR} package) are used, but specified
parameters can be used via the \code{krigeParams} argument. \strong{geoR} has two main
kriging functions: \code{\link[geoR:ksline]{geoR::ksline()}} and \code{\link[geoR:krige.conv]{geoR::krige.conv()}}. The argument
\code{useKsLine} specifies whether to use the \code{\link[geoR:ksline]{geoR::ksline()}} function or not.
}
\examples{
\dontrun{
# You need ggplot2
if (!requireNamespace("ggplot2", quietly = TRUE)) {
  stop("ggplot2 needed for this example to work. Please install it.",
    call. = FALSE)
}

library(ggplot2)

# Use randomized data set for examples (invalid for reaserch).
df <- fgeo.habitat::soil_random
str(df)

# Krige with automated parameters
kriged <- GetKrigedSoil(df, var = "M3Al")
str(kriged)

# Krige with specified parameters - these example params are rather arbitrary
# but are loosely based on what was chosen from the automated kriging
params <- list(
  model = "circular",
  range = 100,
  nugget = 1000,
  sill = 46000,
  kappa = 0.5
)
kriged.2 <- GetKrigedSoil(df, var = "M3Al", krigeParams = params)

# Have a look at the differences
g <- ggplot(kriged$df, aes(x = x, y = y, fill = z)) +
  geom_tile() +
  coord_equal()
g2 <- ggplot(kriged.2$df, aes(x = x, y = y, fill = z)) +
  geom_tile() +
  coord_equal()
g
g2
}
}
\seealso{
\code{\link[geoR:variofit]{geoR::variofit()}}, \code{\link[geoR:variog]{geoR::variog()}}, \code{\link[geoR:as.geodata]{geoR::as.geodata()}},
\code{\link[geoR:ksline]{geoR::ksline()}}, \code{\link[geoR:krige.conv]{geoR::krige.conv()}}, \code{\link[geoR:krige.control]{geoR::krige.control()}}.
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
