% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_lst.R
\name{krig}
\alias{krig}
\alias{GetKrigedSoil}
\title{Krige soil data following the methodology of the John et al. (2007).}
\usage{
krig(soil, var, params = NULL, gridsize = 20,
  plotdim = guess_plotdim(soil), breaks = krig_breaks(2, 320, 30),
  use_ksline = TRUE, quiet = FALSE)

GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = krig_breaks(2, 320, 30),
  useKsLine = TRUE)
}
\arguments{
\item{soil, df.soil}{The data frame with the points, coords specified in the
columns \code{gx}, \code{gy}.}

\item{var}{A character vector giving the name in the soil dataset of the
variable(s) containing soil data to krige. \code{var} must be of length 1 for
\code{GetKrigedSoil()} but can be of length greater than one for \code{krig()}.}

\item{params, krigeParams}{If you want to pass specified kriging parameters;
see \code{\link[=krig_auto_params]{krig_auto_params()}} for each parameter.}

\item{gridsize, gridSize}{Points are kriged to the center points of a grid of
this size.}

\item{plotdim, xSize, ySize}{Numeric vectors giving x and y dimensions of the
plot:
\itemize{
\item \code{plotdim}: Must be of length 2 with the format \code{c(x, y)}.
\item \code{xSize}, \code{ySize}: Each must be of length 1.
}}

\item{breaks}{Breaks/intervals used to calculate the semivariogram, which
only happens if \code{krigeParams = NULL} (default).}

\item{use_ksline, useKsLine}{Use the \code{\link[geoR:ksline]{geoR::ksline()}} function? Use \code{TRUE} to
calculate a "best" semivariogram based on default parameters via
\code{geoR::variogram()}]. Use \code{FALSE} to base calculation on parameters passed
to \code{params}.}

\item{quiet}{Use \code{TRUE} to suppresses messages.}
}
\value{
A list with the following items:
\itemize{
\item \code{df}: Data frame of kriged values (column z) at each grid point (x, y).
\item \code{df.poly}: Data frame of the polynomial surface fitted to the raw data.
\item \code{lambda}: The "lambda" value used in the Box-Cox transform of the raw
data.
\item \code{vg}: A list giving the variogram parameters used for the kriging.
\item \code{vm}: Minimum loss value returned from \code{\link[geoR:variofit]{geoR::variofit()}}.
}
}
\description{
Functions to krige soil data:
\itemize{
\item \code{GetKrigedSoil()} is the original function but it is softly deprecated. Its
result is the same as \code{krig()[[1]]} when run for one soil variable.
\item \code{krig()} some additional convenient features:
\itemize{
\item An interface consistent with other functions of ForestGEO.
\item A a method for \code{summary()}(see examples).
\item Tries to guess \code{plotdim}.
\item Informs the guessed \code{plotdim} and the \code{gridsize} provided.
\item Aallows to  suppress messages.
\item Iterating over multiple soil variables.
\item Has a method for \code{to_df()} (see examples).
}
}

[[1]: R:[1
}
\section{Breaks default}{

The default breaks argument is set to have more points where the exponential
curve rises the most and fewer at large distances. This means that the curve
fitting is not overly biased by points beyond the effective maximum range.
}

\examples{
# Original funciton
# Using automated parameters
auto <- GetKrigedSoil(soil_fake, var = "c")
summary(auto)

# Wrapper produces the same output and add convenient features

# Now using custom parameters (arbitrary but based on automated kriging params)
params <- list(
  model = "circular", range = 100, nugget = 1000, sill = 46000, kappa = 0.5
)

# Also using not one but multiple soil variables
vars <- c("c", "p")
custom <- krig(soil_fake, vars, params = params, quiet = TRUE)
summary(custom)

result <- to_df(custom, item = "df")
head(result)

tail(result)
}
\seealso{
\code{\link[geoR:variofit]{geoR::variofit()}}, \code{\link[geoR:variog]{geoR::variog()}}, \code{\link[geoR:as.geodata]{geoR::as.geodata()}},
\code{\link[geoR:ksline]{geoR::ksline()}}, \code{\link[geoR:krige.conv]{geoR::krige.conv()}}, \code{\link[geoR:krige.control]{geoR::krige.control()}}.

\code{\link[=to_df.krig_lst]{to_df.krig_lst()}}, \code{\link[=summary.krig_lst]{summary.krig_lst()}}.
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
