% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig.R
\name{krig}
\alias{krig}
\alias{GetKrigedSoil}
\title{Krige soil data following the methodology of the John et al. (2007).}
\usage{
krig(soil, var, params = NULL, gridsize = 20, plotdim = c(1000, 500),
  breaks = krig_breaks(2, 320, 30), use_ksline = TRUE)

GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = krig_breaks(2, 320, 30),
  useKsLine = TRUE)
}
\arguments{
\item{soil, df.soil}{The data frame with the points, coords specified in the
columns \code{gx}, \code{gy}.}

\item{var}{A text string giving the variable/column in \code{df.soil} to krige.}

\item{params, krigeParams}{If you want to pass specified kriging parameters;
see \code{\link[=krig_auto_params]{krig_auto_params()}} for each parameter.}

\item{gridsize, gridSize}{Points are kriged to the center points of a grid of
this size.}

\item{plotdim, xSize, ySize}{Numeric vectors giving x and y dimensions of the
plot:
\itemize{
\item \code{plotdim}: Must be of lenght 2 with the format \code{c(x, y)}.
\item \code{xSize}, \code{ySize}: Each must be of lenght 1.
}}

\item{breaks}{Breaks/intervals used to calculate the semivariogram, which
only happens if \code{krigeParams = NULL} (default).}

\item{use_ksline, useKsLine}{Use the \code{\link[geoR:ksline]{geoR::ksline()}} function? Use \code{TRUE} to
calculate a "best" semivariogram based on default parameters via
\code{geoR::variogram()}]. Use \code{FALSE} to base calculation on parameters passed
to \code{params}.}
}
\value{
A list with the following items:
\itemize{
\item \code{df}: Data frame of kriged values (column z) at each grid point (x, y).
\item \code{df.poly}: Data frame of the polynomial surface fitted to the raw data.
\item \code{lambda}: The "lambda" value used in the Box-Cox transform of the raw
data.
\item \code{vg}: A list giving the variogram parameters used for the kriging.
\item \code{vm}: Minimum loss value returned from \code{\link[geoR:variofit]{geoR::variofit()}}.
}
}
\description{
Krige soil data following the methodology of the John et al. (2007).
\code{GetKrigedSoil()} is softly deprecated. Compared to it, \code{krig()} has an
interface that is more consistent with other functions of ForestGEO, and
outputs a list of subclass "krig" -- which has a \code{summary()} method (see
examples).
}
\examples{
\dontrun{
# Example data
soil <- soil_random

# Krige with automated parameters
auto <- krig(soil, var = "m3al")
# For a nice view in RStudio use: View(auto)
summary(auto)

# Custom params (arbitrary but based on automated kriging params)
params <- list(
  model = "circular", range = 100, nugget = 1000, sill = 46000, kappa = 0.5
)
custom <- krig(soil, var = "m3al", params = params)
# For a nice view in RStudio use: View(custom)
summary(custom)

# Compare
if (!requireNamespace("ggplot2", quietly = TRUE)) {
  stop("Install ggplot2 for this section to work.", call. = FALSE)
}
library(ggplot2)

title_auto <- "Automated parameters"
ggplot(auto$df, aes(x = x, y = y, fill = z)) +
  geom_tile() + 
  coord_equal() +
  labs(title = title_auto)
title_custom <- "Custom parameters"

ggplot(custom$df, aes(x = x, y = y, fill = z)) +
  geom_tile() + 
  coord_equal() +
  labs(title = title_custom)
}
}
\seealso{
\code{\link[geoR:variofit]{geoR::variofit()}}, \code{\link[geoR:variog]{geoR::variog()}}, \code{\link[geoR:as.geodata]{geoR::as.geodata()}},
\code{\link[geoR:ksline]{geoR::ksline()}}, \code{\link[geoR:krige.conv]{geoR::krige.conv()}}, \code{\link[geoR:krige.control]{geoR::krige.control()}}.
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
