% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.krig.R
\name{summary.krig}
\alias{summary.krig}
\title{Summary of \code{krig_lst()} results.}
\usage{
\method{summary}{krig}(object, ...)
}
\arguments{
\item{object}{The result of \code{krig_lst()}.}

\item{...}{Arguments passed on to \code{base::summary}
\describe{
  \item{object}{an object for which a summary is desired.}
}}
}
\value{
Prints a cleaner version of \code{str()} and returns its input invisibly.
}
\description{
A cleaner version of \code{str()} for the result of \code{krig_lst()}.
}
\examples{
multiple_results <- krig_lst(soil_fake, c("c", "p"), quiet = TRUE)
summary(multiple_results)

single_result <- multiple_results[[1]]
summary(single_result)
}
