% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.krig.R
\name{summary.krig}
\alias{summary.krig}
\title{Summary of \code{krig()} results.}
\usage{
\method{summary}{krig}(object, ...)
}
\arguments{
\item{object}{The result of \code{krig()}.}

\item{...}{Arguments passed on to \code{base::summary}
\describe{
  \item{object}{an object for which a summary is desired.}
}}
}
\value{
Prints a cleaner version of \code{str()} and returns its input invisibly.
}
\description{
A cleaner version of \code{str()} for the result of \code{krig()}.
}
\examples{
df <- soil_random

res1 <- GetKrigedSoil(df, var = "M3Al")
summary(res1)

res2 <- krig(df, var = "M3Al")
summary(res2)

identical(unclass(res1), unclass(res2))
}
