% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.krig_lst.R
\name{summary.krig_lst}
\alias{summary.krig_lst}
\title{Summary results of \code{krig()}.}
\usage{
\method{summary}{krig_lst}(object, ...)
}
\arguments{
\item{object}{The result of \code{krig()}.}

\item{...}{Arguments passed on to \code{base::summary}
\describe{
  \item{object}{an object for which a summary is desired.}
}}
}
\value{
Prints a cleaner version of \code{str()} and returns its input invisibly.
}
\description{
A cleaner version of \code{str()} for the result of `krig().
}
\examples{
result <- krig(soil_fake, c("c", "p"), quiet = TRUE)
summary(result)
str(result)
}
