% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tt_lst.R
\name{summary.tt_lst}
\alias{summary.tt_lst}
\title{Interpret the output of \code{tt_test()}.}
\usage{
\method{summary}{tt_lst}(object, ...)
}
\arguments{
\item{object}{The output of \code{\link[=tt_test]{tt_test()}}, i.e. an S3 object of class tt_lst.}

\item{...}{Other arguments passed to methods.}
}
\value{
A dataframe.
}
\description{
Interpret the output of \code{tt_test()}.
}
\seealso{
\code{\link[=tt_test]{tt_test()}}.
}
\author{
Daniel Zuleta.
}
