% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torusonesp.all.R
\name{torusonesp.all}
\alias{torusonesp.all}
\title{Torus Translation Test to determine habitat associations of tree species.}
\usage{
torusonesp.all(species, hab.index20, allabund20, plotdim, gridsize)
}
\arguments{
\item{species}{Character sting giving the name of one species.}

\item{hab.index20}{Object giving the habitat designation for each 20x20
quadrat.}

\item{allabund20}{Abundance per quadrat.}

\item{plotdim}{Plot dimensions.}

\item{gridsize}{Grid size. Make sure the gridsize entered here matches the
gridsize on which the habitats are defined and the abundances were
calculated}
}
\value{
A numeric matrix.
}
\description{
Torus Translation Test (TT test) to determine habitat associations of tree
species. Based on code written by Kyle Harms.
}
\details{
You should only try to determine the habitat association for sufficiently
abundant species - in a 50-ha plot, a minimum abundance of 50 trees/species
has been used.
}
\examples{
# Small dataset from Luquillo
census_data <- luquillo_top3_sp
alive_trees <- census_data[census_data$status == "A", ]

habitat_data <- luquillo_habitat
plot_dimensions <- c(320, 500)
grid_size <- 20

abundance_per_quadrat <- abundanceperquad(
  alive_trees,
  plotdim = plot_dimensions,
  gridsize = grid_size,
  type = 'abund'
)$abund

one_species <- unique(census_data$sp)[[1]]
out_one <- torusonesp.all(
  species = one_species,
  hab.index20 = habitat_data,
  allabund20 = abundance_per_quadrat,
  plotdim = plot_dimensions,
  gridsize = grid_size
)
t(out_one)

# Iterate over multiple species
all_species <- unique(census_data$sp)
out_all <- lapply(
  X = all_species,
  FUN = torusonesp.all,
  # Other arguments passed to torusonesp.all
  hab.index20 = habitat_data,
  allabund20 = abundance_per_quadrat,
  plotdim = plot_dimensions,
  gridsize = grid_size
)
# Compact view
t(Reduce(rbind, out_all))
}
\seealso{
Example at \url{https://bookdown.org/fgeocomm/ttt/}.
}
\author{
Sabrina Russo, Daniel Zuletta, Matteo Detto.
}
