% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_test.R
\name{tt_test}
\alias{tt_test}
\title{Torus Translation Test to determine habitat associations of tree species.}
\usage{
tt_test(census, sp, habitat, plotdim = NULL, gridsize = NULL)
}
\arguments{
\item{census}{A dataframe; a ForestGEO \emph{tree} table (see details).}

\item{sp}{Character sting giving any number of species-names.}

\item{habitat}{Object giving the habitat designation for each
plot partition defined by \code{gridsize}. See \code{\link[=fgeo_habitat]{fgeo_habitat()}}.}

\item{plotdim, gridsize}{Plot dimensions and gridsize. If \code{NULL} (default)
they will be guessed, and a message will inform you of the chosen values.
If the guess is wrong, you should provide the correct values manually (and
check that your habitat data is correct).}
}
\value{
A list of matrices. You can summarize the output with \code{\link[=summary]{summary()}} and
convert it to a dataframe with \code{\link[=to_df]{to_df()}}. You can also view the
result with \code{View(your-result)}, and reduce it from a list of matrices to a
single matix with \code{Reduce(rbind, your-result)}. See examples.
}
\description{
Determine habitat-species associations with code developed by Sabrina Russo,
Daniel Zuleta, Matteo Detto, and Kyle Harms.
}
\details{
You should only try to determine the habitat association for sufficiently
abundant species. In a 50-ha plot, a minimum abundance of 50 trees/species
has been used. Also, you should use data of individual trees (i.e. a tree
table, and not a stem table with potentially multiple stems per tree. This
test only makes sense at the population level. We are interested in knowing
whether or not individuals of a species are aggregated on a habitat. Multiple
stems of an individual do not represent population level processes but
individual level processes.
}
\section{Acknowledgments}{

Nestor Engone Obiang, David Kenfack, Jennifer Baltzer, and Rutuja
Chitra-Tarak provided feedback. Daniel Zuleta provided guidance.
}

\section{Interpretation of Output}{

\itemize{
\item \code{N.Hab.1}: Count of stems of the focal species in habitat 1.
\item \code{Gr.Hab.1}: Count of instances the observed relative density of the focal
species on habitat 1 was greater than the relative density based on the TT
habitat map.
\item \code{Ls.Hab.1}: Count of instances the observed relative density of the focal
species on habitat 1 was less than the relative density based on the TT
habitat map.
\item \code{Eq.Hab.1}: Count of instances the observed relative density of the focal
species on habitat 1 was equal to the relative density based on the TT
habitat map.
The sum of the \code{Gr.Hab.x}, \code{Ls.Hab.x}, and \code{Eq.Hab.x} columns for one habitat
equals the number of 20 x20 quads in the plot.
The \code{Rep.Agg.Neut} columns for each habitat indicate whether the species is
significantly repelled (-1), aggregated (1), or neutrally distributed (0) on
the habitat in question.
The probabilities associated with the test for whether these patterns are
statistically significant are in the \code{Obs.Quantile} columns for each habitat.
Note that to calculate the probability for repelled, it is the value given,
but to calculate the probability for aggregated, it is 1- the value given.
}
}

\examples{
library(dplyr)
library(fgeo.data)

# Pick alive trees, of 10 mm or more
census <- filter(luquillo_tree6_random, status == "A", dbh >= 10)

# Pick sufficiently abundant species
pick <- filter(add_count(census, sp), n > 50)
species <- unique(pick$sp)

# Use your habitat data or create if from elevation data
habitat <- fgeo.tool::fgeo_habitat(luquillo_elevation, gridsize = 20, n = 4)

# A list or matrices
tt_lst <- tt_test(census, species, habitat)
tt_lst

# A simple summary to help you interpret the results
summary(tt_lst)

# A combined matrix
Reduce(rbind, tt_lst)

# A dataframe
dfm <- to_df(tt_lst)

# Using dplyr to summarize results by species and distribution
summarize(group_by(dfm, sp, distribution), n = sum(stem_count))
}
\seealso{
\code{\link[=summary.tt_lst]{summary.tt_lst()}}, \code{\link[=to_df]{to_df()}}, \code{\link[=fgeo_habitat]{fgeo_habitat()}}.
}
\author{
Sabrina Russo, Daniel Zuleta, Matteo Detto, and Kyle Harms.
}
